+++ resources/views/pages/home.blade.php
@@ -168,13 +168,13 @@
             Zacznij od zdjęcia tego, co dziś ugotowałeś.
         </x-empty-state>
     @else
-        <div class="stack">
+        <div class="stack" id="lista-wpisow">
             @foreach($posts as $post)
                 <x-post-card :post="$post" />
             @endforeach
         </div>
 
-        <x-show-more :paginator="$posts" />
+        <x-show-more :paginator="$posts" lista="lista-wpisow" />
     @endif
     <section class="marka-start-pomoc" aria-labelledby="start-jak-dziala">
         <p class="nadtytul">Gotowanie łączy</p>